+++ KioChess/Engine/Interfaces/IBitService.cs
@@ -5,7 +5,7 @@ namespace Engine.Interfaces
 {
     public interface IBitService
     {
-        int Count(BitBoard b);
+        byte Count(BitBoard b);
         byte BitScanForward(BitBoard b);
         IEnumerable<byte> BitScan(BitBoard b);
         void GetPositions(BitBoard b, PositionsList positionsList);
